requirements.txt
aiogram==3.22.0
pillow==12.0.0
pydantic==2.11.10
python-dotenv==1.2.1
SQLAlchemy==2.0.44